+++ src/dex/api/muesliswap-api.ts
@@ -52,7 +52,6 @@ export class MuesliSwapApi extends BaseApi {
 
                     liquidityPool.identifier = pool.poolId;
                     liquidityPool.lpToken = new Asset(pool.lpToken.address.policyId, pool.lpToken.address.name);
-                    liquidityPool.totalLpTokens = BigInt(pool.lpToken.amount);
                     liquidityPool.poolFeePercent = Number(pool.poolFee);
 
                     return liquidityPool;