+++ src/dex/wingriders.ts
@@ -129,7 +129,6 @@ export class WingRiders extends BaseDex {
 
         if (lpTokenBalance) {
             liquidityPool.lpToken = lpTokenBalance.asset as Asset;
-            liquidityPool.totalLpTokens = MAX_INT - lpTokenBalance.quantity;
         }
         liquidityPool.poolFeePercent = 0.35;